Fix default bag handling error
When one changed single cluster configuration to multibag configuraion
at some point, buckets and manifests up to that time include no bag
information. These objects should be treated to be contained in "the
default bag", which is the same bag as the master bag in current implementaion.
In source code, the default bag is represented as the atom 'undefined'.